requirements.txt
# Financial Transaction CSV Importer - Dependencies

# Data processing
pandas>=2.0.0

# Database ORM
SQLAlchemy>=2.0.0

# Configuration
PyYAML>=6.0

# CLI display
tabulate>=0.9.0

# Visualization
matplotlib>=3.7.0
altair>=5.0.0

# Web UI (optional, for viewer and analytics dashboard)
streamlit>=1.28.0

# Testing (optional, for development)
pytest>=7.0.0
pytest-cov>=4.0.0

